Reference

SimpleDelimitedFiles.readdlm — Function

readdlm(source, [delim::Char = '\t'], [T::Type = Float64])

Read a delimited numerical matrix from the file named source. The elements of the matrix should be of type T. The end of line delimiter is taken as \n.

For simple usecases, this is a more performant equivalent of readdlm from DelimitedFiles.

Examples

julia> f = tempname();

julia> using DelimitedFiles

julia> writedlm(f, ones(2, 2))

julia> SimpleDelimitedFiles.readdlm(f)
2×2 Matrix{Float64}:
 1.0  1.0
 1.0  1.0

julia> writedlm(f, ones(Int, 2, 2))

julia> SimpleDelimitedFiles.readdlm(f, Int)
2×2 Matrix{Int64}:
 1  1
 1  1

julia> writedlm(f, ones(Int, 2, 2), ',')

julia> SimpleDelimitedFiles.readdlm(f, ',', Int)
2×2 Matrix{Int64}:
 1  1
 1  1